requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asyncio==4.0.0
attrs==25.4.0
bitarray==3.8.0
certifi==2026.2.25
charset-normalizer==3.4.4
ckzg==2.1.6
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
frozenlist==1.8.0
hexbytes==1.3.1
hyperliquid-python-sdk==0.22.0
idna==3.11
magic-filter==1.0.12
msgpack==1.1.2
multidict==6.7.1
parsimonious==0.10.0
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
rlp==4.1.0
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
yarl==1.23.0
